Revert "Job manager"
