+++ client/package.json
@@ -13,7 +13,6 @@
     "@radix-ui/react-label": "^2.1.1",
     "@radix-ui/react-select": "^2.1.4",
     "@radix-ui/react-slot": "^1.1.1",
-    "@radix-ui/react-tabs": "^1.1.2",
     "@radix-ui/react-toast": "^1.2.4",
     "@tremor/react": "^3.18.6",
     "@types/mapbox-gl": "^3.4.1",
