--- client/src/components/jobs/JobsOverview.tsx
@@ -1,60 +1,0 @@
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Job } from "@/types/jobs";
-import { Play, Pause, Check, AlertTriangle } from "lucide-react";
-
-interface JobsOverviewProps {
-  jobs: Job[];
-}
-
-export function JobsOverview({ jobs }: JobsOverviewProps) {
-  const stats = {
-    active: jobs.filter(job => job.status === 'in_progress').length,
-    pending: jobs.filter(job => job.status === 'pending').length,
-    completed: jobs.filter(job => job.status === 'completed').length,
-    failed: jobs.filter(job => job.status === 'failed').length,
-  };
-
-  return (
-    <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-4">
-      <Card>
-        <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-          <CardTitle className="text-sm font-medium">Active Jobs</CardTitle>
-          <Play className="h-4 w-4 text-green-600" />
-        </CardHeader>
-        <CardContent>
-          <div className="text-2xl font-bold">{stats.active}</div>
-        </CardContent>
-      </Card>
-
-      <Card>
-        <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-          <CardTitle className="text-sm font-medium">Pending Jobs</CardTitle>
-          <Pause className="h-4 w-4 text-yellow-600" />
-        </CardHeader>
-        <CardContent>
-          <div className="text-2xl font-bold">{stats.pending}</div>
-        </CardContent>
-      </Card>
-
-      <Card>
-        <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-          <CardTitle className="text-sm font-medium">Completed Jobs</CardTitle>
-          <Check className="h-4 w-4 text-blue-600" />
-        </CardHeader>
-        <CardContent>
-          <div className="text-2xl font-bold">{stats.completed}</div>
-        </CardContent>
-      </Card>
-
-      <Card>
-        <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-          <CardTitle className="text-sm font-medium">Failed Jobs</CardTitle>
-          <AlertTriangle className="h-4 w-4 text-red-600" />
-        </CardHeader>
-        <CardContent>
-          <div className="text-2xl font-bold">{stats.failed}</div>
-        </CardContent>
-      </Card>
-    </div>
-  );
-}
