+++ client/src/hooks/useCustomToast.tsx
@@ -1,4 +1,5 @@
 import { useToast } from './use-toast'
+import { CheckCircle, XCircle } from 'lucide-react'
 
 export function useCustomToast() {
   const { toast } = useToast()
@@ -11,6 +12,7 @@ export function useCustomToast() {
         variant: 'default',
         className: 'bg-gradient-to-r from-green-500 to-green-600 text-white border-none',
         duration: 3000,
+        icon: <CheckCircle className="h-5 w-5" />,
       })
     },
     error: (message: string) => {
@@ -20,6 +22,7 @@ export function useCustomToast() {
         variant: 'destructive',
         className: 'bg-gradient-to-r from-red-500 to-red-600 text-white border-none',
         duration: 3000,
+        icon: <XCircle className="h-5 w-5" />,
       })
     },
   }
