+++ server/Reekon-Take-Home/Robots/Location/GetRobotLocation.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:3001/api/robots/d085f804-af14-434d-ac31-0cd0c0454449/location
+  url: http://localhost:3001/api/robots/{{robotId}}/location
   body: none
   auth: bearer
 }
