+++ server/app.js
@@ -5,7 +5,6 @@ import authRoutes from './routes/authRoutes.js';
 import organizationRoutes from './routes/organizationRoutes.js';
 import robotRoutes from './routes/robotRoutes.js';
 import errorRoutes from './routes/errorRoutes.js';
-import jobRoutes from './routes/jobRoutes.js';
 
 dotenv.config();
 
@@ -20,7 +19,6 @@ app.use('/api/auth', authRoutes);
 app.use('/api/organizations', organizationRoutes);
 app.use('/api/robots', robotRoutes);
 app.use('/api/errors', errorRoutes);
-app.use('/api/jobs', jobRoutes);
 
 
 app.listen(port, () => {
